refactor: move jwt config to application layer
