+++ Source/Endpoints/Iwentys.Endpoints.Api/Startup.cs
@@ -3,6 +3,7 @@ using Iwentys.Infrastructure.Application;
 using Iwentys.Infrastructure.Application.Authorization;
 using Iwentys.Infrastructure.Application.Middlewares;
 using Iwentys.Infrastructure.DataAccess;
+using Iwentys.Infrastructure.DataAccess.Seeding;
 using Iwentys.Integrations.IsuIntegration.Configuration;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Identity;
@@ -23,7 +24,7 @@ namespace Iwentys.Endpoints.Api
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddIwentysIdentity();
+            services.AddIwentysIdentity(Configuration);
             services.EnableExceptional();
             services
                 .AddControllersWithViews(options =>
