+++ Source/Endpoints/Iwentys.Endpoints.Sdk/Iwentys.Endpoints.Sdk.csproj
@@ -7,6 +7,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Connected Services\**" />
+    <EmbeddedResource Remove="Connected Services\**" />
+    <None Remove="Connected Services\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="JsonSubTypes" Version="1.8.0" />
     <PackageReference Include="Polly" Version="7.2.2" />
     <PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
@@ -19,7 +25,6 @@
 
   <ItemGroup>
     <Folder Include="ApiClients\" />
-    <Folder Include="Connected Services\" />
   </ItemGroup>
 
 </Project>
