+++ Source/Infrastructure/Iwentys.Infrastructure.Application/Authorization/JwtApplicationOptions.cs
